pyproject.toml
[project]
name = "requirements"
version = "0.1.0"
description = "dependances"
authors = [
    {name = "hacene"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ultralytics (>=8.3.148,<9.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic (>=2.11.5,<3.0.0)",
    "jose (>=1.0.0,<2.0.0)",
    "duckdb (>=1.3.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
